+++ crates/engine/src/lob/types.rs
@@ -1,3 +1,3 @@
 pub type OrderId = u64;
-pub type Price = u32;
+pub type Price = u64;
 pub type Timestamp = u64;
